pyproject.toml
[project]
name = "my-fastapi-project"
version = "0.1.0"
description = ""
authors = [
    {name = "your name",email = "m.abdullahp454@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymongo (>=4.15.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
